requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
langchain-openai
browser-use